refactor(api): move env config
