+++ api/src/lib/index.ts
@@ -1,4 +1,4 @@
 export * from './logger/logger.module';
 export * from './logger/logger.service';
-export * from './env';
+export * from '../config/env';
 export * from './log';
